refactor: delete extra comments and unused files
