+++ drizzle.config.js
@@ -6,7 +6,7 @@ dotenv.config({ path: path.resolve(__dirname, '.env') });
 
 module.exports = defineConfig({
   out: './src/drizzle',
-  schema: './src/models/schema.js',
+  schema: './src/models/*.js',
   dialect: 'postgresql',
   dbCredentials: {
     url: process.env.DATABASE_URL,
