+++ src/models/productTagSchema.js
@@ -1,6 +1,6 @@
 const { pgTable, serial, integer } = require('drizzle-orm/pg-core');
 
-const { tagsTable } = require('./tagsSchema');
+const { tagsTable } = require('./tagSchema');
 const { productsTable } = require('./productSchema');
 
 const productTagSTable = pgTable('product_tags', {
